refactor: split core validation and runtime modules
Move contract validation out of contract_core, split large semantic/runtime/osctl modules into facades plus focused submodules, and clear hard file-size exemptions.

Keep public type names stable while reducing large-file conflict surfaces across manifest projection, evidence scenarios, semantic graph validation, target executor, and CLI/view tests.

+++ crates/host/kernel/src/supervisor/services/epoll.rs
@@ -51,7 +51,7 @@ impl EpollService {
             .call(&self.create, flags, "epoll_service trapped")
             .map_err(ServiceCallError::Trap)?;
         if raw < 0 {
-            return Err(ServiceCallError::Errno((-raw) as i32));
+            return Err(ServiceCallError::Errno(-raw));
         }
         Ok(raw as u32)
     }
